HZN-1181: Merge ClientFactory and JestClientFactory
