+++ features/internal-plugins-descriptor/pom.xml
@@ -80,13 +80,13 @@
                 <!-- <descriptor>mvn:${project.groupId}/${productName}/${project.version}/xml/features</descriptor> -->
                 <descriptor>mvn:org.opennms.plugins/alarm-change-notifier/${project.version}/xml/features</descriptor>
                 <descriptor>mvn:org.opennms.plugins/opennms-es-rest/${project.version}/xml/features</descriptor>
-                <descriptor>mvn:org.opennms.features/opennms-jest/${project.version}/xml/features</descriptor>
+                <descriptor>mvn:org.opennms.features.jest/opennms-jest/${project.version}/xml/features</descriptor>
               </descriptors>
               <features>
                 <!-- <feature>${productName}/${project.version}</feature> -->
                 <feature>alarm-change-notifier/${project.version}</feature>
                 <feature>opennms-es-rest/${project.version}</feature>
-                <feature>opennms-jest</feature>
+                <feature>opennms-jest/${project.version}</feature>
               </features>
               <repository>${project.build.directory}/temp-features-repo</repository>
             </configuration>
