+++ features/jest/client/src/main/java/org/opennms/plugins/elasticsearch/rest/OnmsJestResult.java
