+++ features/jest/jest-complete-osgi/pom.xml
@@ -1,8 +1,8 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <groupId>org.opennms.plugins</groupId>
-    <artifactId>opennms-es-rest.parent</artifactId>
+    <groupId>org.opennms.features</groupId>
+    <artifactId>org.opennms.features.jest</artifactId>
     <version>22.0.0-SNAPSHOT</version>
   </parent>
