+++ features/opennms-es-rest/pom.xml
@@ -27,7 +27,6 @@
   </properties>
 
   <modules>
-    <module>jest-complete-osgi</module>
     <module>main-module</module>
     <module>product-descriptor</module>
     <module>feature</module>
